# src/BeautifyBaltics.Apps/apps/back-office/src/state/core-api-spec.yaml
openapi: 3.0.4
info:
  title: Beautify Baltics Core API
  description: Beautify Baltics Core API for back-office operations
  version: v1
paths:
...
  '/api/v1/admin/kyc/{masterId}/approve':
    patch:
      tags:
        - AdminKyc
      summary: Approve a master's KYC submission
      operationId: ApproveMasterKyc
      parameters:
        - name: masterId
          in: path
          description: Master identifier
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: No Content
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '422':
          description: Unprocessable Content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationProblemDetails'
  '/api/v1/admin/kyc/{masterId}/reject':
    patch:
      tags:
        - AdminKyc
      summary: Reject a master's KYC submission with a required reason
      operationId: RejectMasterKyc
      parameters:
        - name: masterId
          in: path
          description: Master identifier
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        description: Rejection reason
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KycRejectRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/KycRejectRequest'
      responses:
        '204':
          description: No Content
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '422':
          description: Unprocessable Content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationProblemDetails'
...
components:
  schemas:
...
    KycRejectRequest:
      required:
        - reason
      type: object
      properties:
        reason:
          minLength: 1
          type: string
      additionalProperties: false
...
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: integer
          format: int32
          nullable: true
        detail:
          type: string
          nullable: true
        instance:
          type: string
          nullable: true
      additionalProperties: { }
...
    ValidationProblemDetails:
      type: object
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: integer
          format: int32
          nullable: true
        detail:
          type: string
          nullable: true
        instance:
          type: string
          nullable: true
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          nullable: true
      additionalProperties: { }